init cleanup of user migration
