+++ src/Auth/Auth.php
@@ -4,15 +4,15 @@ declare(strict_types=1);
 
 namespace Canvas\Auth;
 
-use Canvas\Models\Users;
+use Baka\Auth\UserProvider;
 use Baka\Contracts\Auth\UserInterface;
 use Baka\Exception\AuthException;
 use Baka\Hashing\Keys;
 use Baka\Hashing\Password;
 use Baka\Support\Random;
+use Canvas\Models\Users;
 use Exception;
 use stdClass;
-use Baka\Auth\UserProvider;
 
 class Auth
 {
@@ -64,10 +64,10 @@ class Auth
     }
 
     /**
-    * Create a new user.
-    *
-    * @return Users
-    */
+     * Create a new user.
+     *
+     * @return Users
+     */
     public static function signUp(UserInterface $user) : UserInterface
     {
         $user->sex = 'U';
@@ -108,7 +108,7 @@ class Auth
      *
      * @throws Exception
      *
-     * @return void
+     * @return bool
      */
     protected static function loginAttemptsValidation(UserInterface $user) : bool
     {
