+++ src/Models/Users.php
@@ -3,8 +3,8 @@ declare(strict_types=1);
 
 namespace Canvas\Models;
 
-use Canvas\Auth\Models\Users as BakUser;
 use Baka\Cashier\Billable;
+use Baka\Contracts\Auth\AuthTokenTrait;
 use Baka\Contracts\Auth\UserInterface;
 use Baka\Contracts\Database\HashTableTrait;
 use Baka\Contracts\EventsManager\EventManagerAwareTrait;
@@ -13,6 +13,7 @@ use Baka\Hashing\Keys;
 use Baka\Hashing\Password;
 use Baka\Validations\PasswordValidation;
 use Canvas\Auth\App as AppAuth;
+use Canvas\Auth\Models\Users as BakUser;
 use Canvas\Contracts\FileSystemModelTrait;
 use Canvas\Contracts\PermissionsTrait;
 use Canvas\Contracts\SubscriptionPlanLimitTrait;
@@ -26,7 +27,7 @@ use Phalcon\Validation\Validator\PresenceOf;
 use Phalcon\Validation\Validator\Uniqueness;
 use ReflectionClass;
 
-class Users extends BakUser implements UserInterface
+class Users extends AbstractModel implements UserInterface
 {
     use PermissionsTrait;
     use Billable;
@@ -35,7 +36,56 @@ class Users extends BakUser implements UserInterface
     use HashTableTrait;
     use NotifiableTrait;
     use EventManagerAwareTrait;
+    use AuthTokenTrait;
 
+    /**
+     * Constant for anonymous user.
+     */
+    const ANONYMOUS = '-1';
+
+    public ?string $email = null;
+    public ?string $password = null;
+    public ?string $firstname = null;
+    public ?string $lastname = null;
+    public ?string $displayname = null;
+    public ?string $registered = null;
+    public ?string $lastvisit = null;
+    public int $default_company = 0;
+    public ?string $defaultCompanyName = null;
+    public ?string $dob = null;
+    public ?string $sex = null;
+    public ?string $description = null;
+    public ?string $phone_number = null;
+    public ?string $cell_phone_number = null;
+    public ?string $timezone = null;
+    public ?int $city_id = 0;
+    public ?int $state_id = 0;
+    public ?int $country_id = 0;
+    public int $welcome = 0;
+    public int $user_active = 0;
+    public ?string $user_activation_key = null;
+    public ?string $user_activation_email = null;
+    public ?string $profile_header = '';
+    public bool $loggedIn = false;
+    public ?string $location = null;
+    public string $interest = '';
+    public int $profile_privacy = 0;
+    public ?string $user_activation_forgot = null;
+    public ?string $language = null;
+    public string $session_id = '';
+    public string $session_key = '';
+    public ?string $banned = null;
+    public ?int $user_last_login_try = 0;
+    public int $user_level = 0;
+    public static string $locale = 'ja_jp';
+
+    /**
+     * @deprecated with filesystem
+     */
+    public ?string $profile_image = null;
+    public ?string $profile_image_mobile = null;
+    public ?string $profile_remote_image = null;
+    public ?string $profile_image_thumb = ' ';
     public int $default_company_branch = 0;
     public ?int $roles_id = null;
     public ?string $stripe_id = null;
@@ -251,6 +301,172 @@ class Users extends BakUser implements UserInterface
         );
 
         return $this->validate($validator);
+    }
+
+    /**
+     * get the user by its Id, we can specify the cache if we want to
+     * we only get result if the user is active.
+     *
+     * @param int $userId
+     * @param bool $cache
+     *
+     * @return UserInterface
+     */
+    public static function getById($id, $cache = false) : UserInterface
+    {
+        $options = null;
+        $key = 'userInfo_' . $id;
+
+        if ($cache) {
+            $options = ['cache' => ['lifetime' => 3600, 'key' => $key]];
+        }
+
+        return self::findFirstOrFail([
+            'conditions' => 'id = ?0 and is_deleted = 0',
+            'bind' => [$id]
+        ]);
+    }
+
+    /**
+     * is the user active?
+     *
+     * @return bool
+     */
+    public function isActive() : bool
+    {
+        return $this->user_active;
+    }
+
+    /**
+     * get user by there email address.
+     *
+     * @return User
+     */
+    public static function getByEmail(string $email) : UserInterface
+    {
+        $user = self::findFirst([
+            'conditions' => 'email = ?0 and is_deleted = 0',
+            'bind' => [$email]
+        ]);
+
+        if (!$user) {
+            throw new Exception('No User Found');
+        }
+
+        return $user;
+    }
+
+    /**
+     * get user nickname.
+     *
+     * @return string
+     */
+    public function getDisplayName() : string
+    {
+        return strtolower($this->displayname);
+    }
+
+    /**
+     * get user email.
+     *
+     * @return string
+     */
+    public function getEmail() : ?string
+    {
+        return $this->email;
+    }
+
+    /**
+     * is the user logged in?
+     *
+     * @return bool
+     */
+    public function isLoggedIn() : bool
+    {
+        return $this->loggedIn;
+    }
+
+    /**
+     * Is Anonymous user.
+     *
+     * @return bool
+     */
+    public function isAnonymous() : bool
+    {
+        return (int) $this->getId() == self::ANONYMOUS;
+    }
+
+    /**
+     * get the user sex, not get sex from the user :P.
+     *
+     * @return string
+     */
+    public function getSex() : string
+    {
+        if ($this->sex == 'M') {
+            return _('Male');
+        } elseif ($this->sex == 'F') {
+            return _('Female');
+        } else {
+            return _('Undefined');
+        }
+    }
+
+    /**
+     * Log a user out of the system.
+     *
+     * @return bool
+     */
+    public function logOut() : bool
+    {
+        $session = new Sessions();
+        $session->end($this);
+
+        return true;
+    }
+
+    /**
+     * Clean the user session from the system.
+     *
+     * @return true
+     */
+    public function cleanSession() : bool
+    {
+        $session = new Sessions();
+        $session->end($this);
+
+        return true;
+    }
+
+    /**
+     * get the user session id.
+     *
+     * @return string
+     */
+    public function getSessionId() : string
+    {
+        //if its empty get it from the relationship, else get it from the property
+        return empty($this->session_id) ? $this->getSession(['order' => 'time desc'])->session_id : $this->session_id;
+    }
+
+    /**
+     * get the user language.
+     *
+     * @return string
+     */
+    public function getLanguage() : ? string
+    {
+        return $this->language;
+    }
+
+    /**
+     * Determine if a user is banned.
+     *
+     * @return bool
+     */
+    public function isBanned() : bool
+    {
+        return !$this->isActive() && $this->banned === 'Y';
     }
 
     /**
